refactor: Add support for setting analog stick values in AIControllerMappings.json and AICLib.py
This commit adds support for setting analog stick values in the AIControllerMappings.json file and the AICLib.py file. The "ANALOG_STICKS" key is added to the AIControllerMappings.json file with the corresponding value "set_analog_stick_values". In the AICLib.py file, a new method called "set_analog_stick_values" is implemented, which takes four parameters (x, y, rx, ry) and sets the corresponding values in the AIController object. This change allows for controlling analog sticks in the AIController module.