refactor: Update .gitignore and package.json dependencies
This commit updates the .gitignore file to exclude the "node_modules" directory and includes the "node_modules" directory in the .gitignore file of the "Website" directory. Additionally, the package.json file in the "Website" directory is updated to include new dependencies: "autoprefixer", "react-router-dom", "run", and "tailwindcss". These updates improve the project's version control and ensure that the necessary dependencies are included for proper functionality.

Co-authored-by: Emil Deßloch <Limonadeundco@users.noreply.github.com>