feat: Add endpoint for retrieving last log data
This commit adds a new endpoint "/log/data/<int:lines>" to retrieve the last <int> lines of log data. The endpoint reads the most recent log file from the "LOGS" folder and returns the specified number of lines. This enhancement improves the functionality of the log feature in the web server.