feat: Update default IP address and port in RPI_COM classes
This commit updates the default IP address and port in the RPI_COM classes. The `Messenger` class now sets the default IP address to "127.0.0.1" and the default port to 11111. Similarly, the `MessageReceiver` class sets the default IP address to "127.0.0.1" and the default port to 11111. This change improves the default configuration for communication between the Raspberry Pis.